WIP: Scope/Cairo/Plot mostly works
stubbed out plotting of axes and cursor for now.
scopeCairoNew calls scopeNew which no longer exists.  Not sure what to replace
it with.  There is no data source at this point, so that part of the logic
(creating a new source and connecting it) doesn't need to happen yet.

Also, I think that Cairo/Plot should call toGtkCoords somewhere.